backend/medical-research-assistant/requirements.txt
google-cloud-aiplatform==1.44.0 #1.48.0
langchain==0.1.15 #0.1.16
langchain-google-vertexai==0.1.2 #==1.0.1
langchain-community==0.0.32 #0.0.34
# pgvector
# psycopg2-binary
cloud-sql-python-connector[pg8000]==1.8.0
Flask==3.0.3
Flask-Cors==5.0.0